vp: linux-vp: added single core variants
Newly added variants:
 * linux-sc-vp:   RV64 with single worker core
 * linux32-sc-vp: RV32 with single worker core

The motivation for this is mainly to speed up some experiments.
For many experiments, a single core is sufficient and simulating a
single core is much faster than simulating multiple cores.

The new variants were introduced in the same way as it was done for the
merge of linux-vp and linux32-vp: The distinction between is controlled
by a define (NUM_CORES) which is set in the corresponding CMakeLists.

It was also considered to make the number of cores configurable by a
command-line parameter. However, because of the structure of the vp
(templates)

It was also considered to make the number of cores configurable via a
command line parameter. However, the VP and SystemC make much use of
template classes with the number of cores as a parameter. This could
perhaps be worked around, but especially with sockets and bindings
this quickly becomes messy and possibly inefficient.

--
manfred.schlaegl@jku.at
Institute for Complex Systems, JKU Linz
https://ics.jku.at/