Added a light weight real-time clint and use it for linux (real wall clock time)
Added lightweight real-time clint lwrt_clint with exactly the same behavior
as clint.h, but mtime is based on real(host) wall clock time instead of
simulation time.

Using lwrt_clint for the linux-vp means, that the wall-clock of the
Linux system running within the simulation is synchronized with the
world outside the simulation. This ensures a valid timings for
interaction (e.g. user I/O) or communication (e.g. networking) with the
real world.