VNCSimpleInput: enable TRACK_CHANGED_AREA optimization
With TRACK_CHANGED_AREA enabled, we don't only track if there was a
change, but also which area was changed. This comes with an overhead
for each change (write to intermediate frame buffer), but reduces the
amount of data to copy to the RFB frame buffer in updateScreen.

Especially when small changes are dominating this leads to improved
performance. (e.g. prboom 352x250)