lwrt_clint: fixed timer interrupt spurious triggers (performance)
NOTE: A similar problem exists in the original clint implementation.
This is not addressed here!

With the current implementation trigger_timer_interrupt(false) is called
very frequently. The problem with this: This does not only reset the
pending flag, but also leads to a resume of WFI in the ISS (This
behavior is also questionable and must also be investigated in the future).
Here not only one ISS, but every core is affected.

The solution: trigger_timer_interrupt(false) is called only for one
specific core, when its corresponding mtimecmp is written.
Matching
"The RISC-V Instruction Set Manual
Volume II: Privileged Architecture
Document Version 20211203" - Page 44ff
-> Much less unnecessary resumes from WFI